refactor: move Claude schemas from src/api/claude to src/claude
