+++ whatever-code/src/api/definitions/index.ts
@@ -1,5 +1,5 @@
 import { ClaudeRpcs } from "./claude.js";
 
-export * as Claude from "../claude/index.js";
+export * as Claude from "../../claude/index.js";
 
 export const ApiRpcs = ClaudeRpcs.merge();
